Move validators to dedicated file
