+++ output.go
@@ -704,36 +704,10 @@ func OutputsSyntacticalImplicitAccountCreationAddress() OutputsSyntacticalValida
 	}
 }
 
-type ElementValidationFunc[T any] func(index int, next T) error
-
-func LexicalOrderAndUniqueness[T constraints.Comparable[T]]() ElementValidationFunc[T] {
-	var prev *T
-	var prevIndex int
-
-	return func(index int, next T) error {
-		if prev == nil {
-			prev = &next
-			prevIndex = index
-		} else {
-			switch (*prev).Compare(next) {
-			case 1:
-				return ierrors.Wrapf(ErrArrayValidationOrderViolatesLexicalOrder, "element %d should have been before element %d", index, prevIndex)
-			case 0:
-				return ierrors.Wrapf(ErrArrayValidationViolatesUniqueness, "element %d and element %d are duplicates", index, prevIndex)
-			}
-
-			prev = &next
-			prevIndex = index
-		}
-
-		return nil
-	}
-}
-
 // TODO
 func OutputsSyntacticalUnlockConditionLexicalOrderAndUniqueness() OutputsSyntacticalValidationFunc {
 	return func(index int, output Output) error {
-		elementValidationFunc := LexicalOrderAndUniqueness[UnlockCondition]()
+		elementValidationFunc := LexicalOrderAndUniquenessValidator[UnlockCondition]()
 		switch typedOutput := output.(type) {
 		case *BasicOutput:
 			for idx, uc := range typedOutput.UnlockConditions {
@@ -767,8 +741,8 @@ func OutputsSyntacticalUnlockConditionLexicalOrderAndUniqueness() OutputsSyntact
 // TODO
 func OutputsSyntacticalFeaturesLexicalOrderAndUniqueness() OutputsSyntacticalValidationFunc {
 	return func(index int, output Output) error {
-		featureValidationFunc := LexicalOrderAndUniqueness[Feature]()
-		immutableFeatureValidationFunc := LexicalOrderAndUniqueness[Feature]()
+		featureValidationFunc := LexicalOrderAndUniquenessValidator[Feature]()
+		immutableFeatureValidationFunc := LexicalOrderAndUniquenessValidator[Feature]()
 
 		switch typedOutput := output.(type) {
 		case *BasicOutput:
