+++ syntax_validator.go
@@ -1,0 +1,47 @@
+package iotago
+
+import (
+	"github.com/iotaledger/hive.go/constraints"
+	"github.com/iotaledger/hive.go/ierrors"
+)
+
+type ElementValidationFunc[T any] func(index int, next T) error
+
+func LexicalOrderAndUniquenessValidator[T constraints.Comparable[T]]() ElementValidationFunc[T] {
+	var prev *T
+	var prevIndex int
+
+	return func(index int, next T) error {
+		if prev == nil {
+			prev = &next
+			prevIndex = index
+		} else {
+			switch (*prev).Compare(next) {
+			case 1:
+				return ierrors.Wrapf(ErrArrayValidationOrderViolatesLexicalOrder, "element %d should have been before element %d", index, prevIndex)
+			case 0:
+				return ierrors.Wrapf(ErrArrayValidationViolatesUniqueness, "element %d and element %d are duplicates", index, prevIndex)
+			}
+
+			prev = &next
+			prevIndex = index
+		}
+
+		return nil
+	}
+}
+
+// TODO: Extend doc.
+// Helper function to validate a slice syntactically.
+func SyntacticSliceValidator[T any](
+	slice []T,
+	elementValidationFunc ElementValidationFunc[T],
+) error {
+	for i, element := range slice {
+		if err := elementValidationFunc(i, element); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
